Abstracted out Okta registration from the /invite endpoint to middleware.
